Cambios comentarios PR

+++ src/main/java/org/emprenApp/detalle_pedido/infrastructure/response/DetallePedidoResponse.java
@@ -3,13 +3,15 @@ package org.emprenApp.detalle_pedido.infrastructure.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class DetallePedidoResponse {
     private Long id;
     private Long pedidoId;
     private Long productoId;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
     private Integer cantidad;
-    private Double subtotal;
+    private BigDecimal subtotal;
 }
